requirements.txt
asyncpg==0.21.0
fastapi==0.61.1
locust==1.2.3
tortoise-orm==0.16.14
uvicorn==0.11.8
